+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added support for loading and previewing root motion translation for animations.
 * Added `merge-databases` command to xc3_shader for creating combined databases.
+* Added support to xc3_gltf for loading `monolib/shader` images.
 
 ### Changed
 * Optimized shader database size.